requirements.txt
bcrypt~=4.0.1
PyJWT~=2.6.0
Flask~=2.2.2
Werkzeug~=2.2.2
mariadb~=1.1.5.post3
elasticsearch~=8.6.0
python-dotenv~=0.21.0